remove sorting functionality for authors
