+++ package.json
@@ -26,7 +26,6 @@
     "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.4",
-    "react-sortable-hoc": "^1.11.0",
     "react-truncate": "^2.4.0",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
