+++ src/pages/registration/ContributorsPanel.tsx
@@ -3,11 +3,11 @@ import { useTranslation } from 'react-i18next';
 import { FormHelperText, Typography } from '@material-ui/core';
 import { FormikProps, useFormikContext, FieldArray, ErrorMessage, FieldArrayRenderProps } from 'formik';
 import Card from '../../components/Card';
-import SortableTable from './contributors_tab/components/SortableTable';
 import { Registration } from '../../types/registration.types';
 import { ContributorFieldNames } from '../../types/publicationFieldNames';
 import { touchedContributorTabFields } from '../../utils/formik-helpers';
 import { PanelProps } from './RegistrationFormContent';
+import Authors from './contributors_tab/Authors';
 
 const ContributorsPanel: FC<PanelProps> = ({ setTouchedFields }) => {
   const { t } = useTranslation('registration');
@@ -35,7 +35,7 @@ const ContributorsPanel: FC<PanelProps> = ({ setTouchedFields }) => {
       <FieldArray name={ContributorFieldNames.CONTRIBUTORS}>
         {({ push, replace, name }: FieldArrayRenderProps) => (
           <>
-            <SortableTable push={push} replace={replace} />
+            <Authors push={push} replace={replace} />
             {contributors.length === 0 && (
               <FormHelperText error>
                 <ErrorMessage name={name} />
